+++ lib/screens/dashboard/dashboard_controller.dart
@@ -3,8 +3,11 @@ import 'package:vase/colors.dart';
 import 'package:vase/const.dart';
 import 'package:vase/controllers/db_controller.dart';
 import 'package:vase/screens/dashboard/dashboard_model.dart';
+import 'package:vase/utils.dart';
 
 class DashboardController extends GetxController {
+  final DateTime currentDate;
+  DashboardController(this.currentDate);
   RxList<Sector> sectors = RxList.empty();
   DbController dbController = Get.find();
 
@@ -18,7 +21,10 @@ class DashboardController extends GetxController {
     var transList = await dbController.db.rawQuery(
       '''SELECT SUM(${Const.trans}.amount) AS total , count(${Const.categories}.category_name) AS share , ${Const.categories}.category_name  from ${Const.trans} LEFT JOIN ${Const.categories}
           on ${Const.trans}.category_id = ${Const.categories}.id
+          WHERE ${Const.trans}.created_at BETWEEN ${Utils.getFirstDate(currentDate)} AND ${Utils.getLastDate(currentDate)}
           GROUP BY ${Const.categories}.category_name
+          ORDER BY share DESC
+          LIMIT 5
       ''',
     );
     for (int i = 0; i < transList.length; i++) {
